Split benchmark for fused operators
